+++ packages/util/src/hex/toBn.ts
@@ -4,7 +4,6 @@
 import type { ToBnOptions } from '../types';
 
 import { BN } from '../bn/bn';
-import { isBoolean } from '../is/boolean';
 import { hexStripPrefix } from './stripPrefix';
 
 /**
@@ -25,19 +24,11 @@ import { hexStripPrefix } from './stripPrefix';
  * hexToBn('0x123480001f'); // => BN(0x123480001f)
  * ```
  */
-function hexToBn (value?: string | null, options?: ToBnOptions): BN;
-/** @deprecated Use hexToBn (value?: string | null, options?: ToBnOptions) */
-function hexToBn (value?: string | null, options?: boolean): BN;
-/** @deprecated Use hexToBn (value?: string | null, options?: ToBnOptions) */
-function hexToBn (value?: string | null, options: ToBnOptions | boolean = {}): BN {
+function hexToBn (value?: string | null, { isLe = false, isNegative = false }: ToBnOptions = {}): BN {
   if (!value || value === '0x') {
     return new BN(0);
   }
 
-  // For hex, default to BE
-  const { isLe = false, isNegative = false } = isBoolean(options)
-    ? { isLe: options }
-    : options;
   const stripped = hexStripPrefix(value);
   const bn = new BN(stripped, 16, isLe ? 'le' : 'be');